Adding files for Feature1 (Customer)
Have added the files that have been edited slightly for the Customer Table and application to work.
Made the Interface a little prettier.
Edited the Jacaco version to 0.8.6 for compat issues with Java 16.
Edited the schema.
I believe at the moment we need to import the schema for it to work.
Will check this as we go to Feature 2. (Items Table)